requirements.txt
Django>=5.2.2
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
python-docx>=0.8.11
PyPDF2>=3.0.0
scikit-learn>=1.4.0
nltk>=3.8.1
joblib>=1.3.2 